Use std::size_t for filling nodal positions

+++ src/mesh/nodal_coordinates.cpp
@@ -2,7 +2,6 @@
 #include "nodal_coordinates.hpp"
 
 #include "Exceptions.hpp"
-
 #include "io/json.hpp"
 
 namespace neon
@@ -22,7 +21,7 @@ nodal_coordinates::nodal_coordinates(json const& mesh_file)
 
     X.resize(3, nodes);
 
-    for (std::int64_t node{0}; node < nodes; ++node)
+    for (std::size_t node{0}; node < nodes; ++node)
     {
         for (auto i = 0; i < 3; ++i)
         {
